centralise header and shift chatbot to components
