+++ frontend/iQMA-Skills-Builder/components/LogoHeader.tsx
@@ -1,0 +1,32 @@
+import { ActivityIndicator, Image, StyleSheet, Text, View } from 'react-native';
+
+const LogoHeader = () => {
+    return (
+    <View style= {styles.container}>
+        <Image source={{uri: 'https://reactjs.org/logo-og.png'}}
+        style={styles.logo} />
+        <Text style={styles.title}>IQMA Skills Builder</Text>
+    </View>
+    );
+  }
+
+
+  const styles = StyleSheet.create({
+    container: {
+        flexDirection: 'row',
+        flex: 1,
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+    logo: {
+        width: 25,
+        height: 25,
+        marginRight: 10,
+    },
+    title: {
+        fontSize: 18,
+        color: '#fff', 
+    },
+});
+
+export default LogoHeader;
